refactoring and separate components
